feat(runtime-core): share pushed health hub
