--- packages/runtime-core/src/index.test.ts
@@ -1,2519 +1,0 @@
-import { describe, expect, it } from "@effect/vitest";
-import type { ColumnLiveViewEngineHealth } from "@effect-view-server/column-live-view-engine";
-import { createColumnLiveViewEngineInternal } from "@effect-view-server/column-live-view-engine/internal";
-import {
-  defineViewServerConfig,
-  kafka,
-  type ViewServerRuntimeError,
-} from "@effect-view-server/config";
-import { grpcSourceMarkers } from "@effect-view-server/config/internal";
-import { Clock, Deferred, Effect, Fiber, Option, Queue, Schema, Stream, Tracer } from "effect";
-import { TestClock } from "effect/testing";
-import { AtomRef } from "effect/unstable/reactivity";
-import {
-  healthFromEngine,
-  makeCoalescedHealthReader,
-  makeHealthRefreshScheduler,
-  readHealth,
-} from "./health";
-import { makeViewServerRuntimeCoreInternal } from "./internal";
-import { createViewServerRuntimeCore, makeViewServerRuntimeCore } from "./index";
-import { makeRuntimeCoreLiveClient } from "./live-client";
-
-const Order = Schema.Struct({
-  id: Schema.String,
-  customerId: Schema.String,
-  status: Schema.Literals(["open", "closed", "cancelled"]),
-  price: Schema.Number,
-  region: Schema.String,
-  updatedAt: Schema.Number,
-});
-
-const viewServer = defineViewServerConfig({
-  topics: {
-    orders: {
-      schema: Order,
-      key: "id",
-    },
-  },
-});
-
-const leasedViewServer = defineViewServerConfig({
-  topics: {
-    orders: {
-      schema: Order,
-      key: "id",
-      grpcSource: grpcSourceMarkers.leased({
-        routeBy: ["region", "status"],
-      }),
-    },
-  },
-});
-
-const leasedGrpcSourceViewServer = defineViewServerConfig({
-  topics: {
-    orders: {
-      schema: Order,
-      key: "id",
-      grpcSource: grpcSourceMarkers.leased({
-        routeBy: ["region", "status"],
-      }),
-    },
-  },
-});
-
-const materializedGrpcSourceViewServer = defineViewServerConfig({
-  topics: {
-    orders: {
-      schema: Order,
-      key: "id",
-      grpcSource: grpcSourceMarkers.materialized(),
-    },
-  },
-});
-
-const kafkaOwnedViewServer = defineViewServerConfig({
-  kafka: {
-    usa: "localhost:9092",
-  },
-  topics: {
-    orders: {
-      schema: Order,
-      key: "id",
-      kafkaSource: kafka.source({
-        topic: "orders-source",
-        regions: ["usa"],
-        value: kafka.json(() => Schema.toCodecJson(Order)),
-        key: kafka.stringKey(),
-        rowKey: ({ key }) => key,
-        map: ({ value }) => ({
-          customerId: value.customerId,
-          status: value.status,
-          price: value.price,
-          region: value.region,
-          updatedAt: value.updatedAt,
-        }),
-      }),
-    },
-  },
-});
-
-type OrderRow = typeof Order.Type;
-type Topics = typeof viewServer.topics;
-
-type RecordedSpan = {
-  readonly attributes: ReadonlyArray<readonly [string, unknown]>;
-  readonly name: string;
-  readonly parentName: string | null;
-  readonly parentSpanId: string | null;
-  readonly spanId: string;
-  readonly traceId: string;
-};
-
-const order = (id: string, price: number): OrderRow => ({
-  id,
-  customerId: `customer-${id}`,
-  status: "open",
-  price,
-  region: "usa",
-  updatedAt: price,
-});
-
-const refreshFailed: ViewServerRuntimeError = {
-  _tag: "ViewServerRuntimeError",
-  code: "RuntimeUnavailable",
-  message: "Health refresh failed.",
-};
-
-const publicLeasedRuntimeAccessError = {
-  _tag: "ViewServerRuntimeError",
-  code: "UnsupportedQuery",
-  topic: "orders",
-  message:
-    "Leased gRPC topics do not support direct runtime mutations, one-shot snapshots, or runtime-core subscriptions; use the runtime gRPC lease manager so it owns lease lifecycle.",
-} satisfies ViewServerRuntimeError;
-
-const publicSourceOwnedRuntimeMutationError = {
-  _tag: "ViewServerRuntimeError",
-  code: "UnsupportedQuery",
-  topic: "orders",
-  message:
-    "Source-owned topics do not support direct runtime mutations; publish through the configured Kafka/gRPC source or use an externally-published topic.",
-} satisfies ViewServerRuntimeError;
-
-const publicSourceOwnedRuntimeResetError = {
-  _tag: "ViewServerRuntimeError",
-  code: "UnsupportedQuery",
-  message:
-    "Source-owned topics do not support direct runtime reset; close the runtime or reset source-free topics through their owner.",
-} satisfies ViewServerRuntimeError;
-
-const stableAttributes = (
-  attributes: ReadonlyMap<string, unknown>,
-): ReadonlyArray<readonly [string, unknown]> =>
-  Array.from(attributes.entries()).sort(([left], [right]) => left.localeCompare(right));
-
-const spanName = (span: Tracer.AnySpan): string => (span._tag === "Span" ? span.name : span.spanId);
-
-const makeRecordingTracer = (): {
-  readonly spans: Array<RecordedSpan>;
-  readonly tracer: Tracer.Tracer;
-} => {
-  const spans: Array<RecordedSpan> = [];
-  let nextSpanId = 0;
-  const nextId = (): string => {
-    nextSpanId += 1;
-    return String(nextSpanId);
-  };
-  const tracer = Tracer.make({
-    span: (options): Tracer.Span => {
-      const id = nextId();
-      const attributes = new Map<string, unknown>();
-      const links = Array.from(options.links);
-      let status: Tracer.SpanStatus = {
-        _tag: "Started",
-        startTime: options.startTime,
-      };
-      const span: Tracer.Span = {
-        _tag: "Span",
-        annotations: options.annotations,
-        attribute: (key, value) => {
-          attributes.set(key, value);
-        },
-        attributes,
-        end: (endTime, exit) => {
-          const parent = Option.getOrNull(options.parent);
-          status = {
-            _tag: "Ended",
-            endTime,
-            exit,
-            startTime: status.startTime,
-          };
-          spans.push({
-            attributes: stableAttributes(attributes),
-            name: options.name,
-            parentName: parent === null ? null : spanName(parent),
-            parentSpanId: parent === null ? null : parent.spanId,
-            spanId: span.spanId,
-            traceId: span.traceId,
-          });
-        },
-        event: () => {},
-        addLinks: (newLinks) => {
-          links.push(...newLinks);
-        },
-        get status() {
-          return status;
-        },
-        kind: options.kind,
-        links,
-        name: options.name,
-        parent: options.parent,
-        sampled: options.sampled,
-        spanId: `span-${id}`,
-        traceId: Option.match(options.parent, {
-          onNone: () => `trace-${id}`,
-          onSome: (parent) => parent.traceId,
-        }),
-      };
-      return span;
-    },
-  });
-  return { spans, tracer };
-};
-
-const engineHealth = (
-  status: "ready" | "stopping",
-  rowCount: number,
-): ColumnLiveViewEngineHealth<Topics> => ({
-  status,
-  version: status === "ready" ? 1 : 2,
-  topics: {
-    orders: {
-      status: status === "ready" ? "ready" : "degraded",
-      rowCount,
-      liveRowCount: rowCount,
-      deletedRowCount: 0,
-      version: status === "ready" ? 1 : 2,
-      lastMutationAt: null,
-      mutationsPerSecond: 0,
-      rowsPerSecond: 0,
-      pendingMutationBatches: 0,
-      activeFallbackGroupedViews: 0,
-      activeIncrementalGroupedViews: 0,
-      activeViews: 0,
-      groupedFullEvaluationCount: 0,
-      groupedPatchedEvaluationCount: 0,
-      activeSubscriptions: 0,
-      queuedEvents: 0,
-      maxQueueDepth: 0,
-      backpressureEvents: 0,
-      memoryBytes: 0,
-      tombstoneCount: 0,
-      compactionPending: false,
-    },
-  },
-  activeSubscriptions: 0,
-  queuedEvents: 0,
-  maxQueueDepth: 0,
-  backpressureEvents: 0,
-});
-
-describe("@effect-view-server/runtime-core", () => {
-  it.effect("reports elapsed runtime uptime and passes refresh time to health overlays", () =>
-    Effect.gen(function* () {
-      yield* TestClock.setTime(10_000);
-      const overlayTimes: Array<number> = [];
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {
-        healthOverlay: (health, nowMillis) => {
-          overlayTimes.push(nowMillis);
-          return health;
-        },
-      });
-
-      expect(runtimeCore.liveClient.health.value.uptimeMs).toBe(0);
-      expect(overlayTimes).toStrictEqual([10_000]);
-
-      yield* TestClock.adjust("2500 millis");
-      const refreshedHealth = yield* runtimeCore.refreshHealth;
-
-      expect(refreshedHealth.uptimeMs).toBe(2_500);
-      expect(overlayTimes).toStrictEqual([10_000, 12_500]);
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("clamps runtime uptime when the clock moves before the runtime start", () =>
-    Effect.gen(function* () {
-      yield* TestClock.setTime(10_000);
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {});
-
-      yield* TestClock.setTime(9_000);
-      const refreshedHealth = yield* runtimeCore.refreshHealth;
-
-      expect(refreshedHealth.uptimeMs).toBe(0);
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("uses monotonic time for uptime when wall time moves backward", () =>
-    Effect.gen(function* () {
-      let wallMillis = 10_000;
-      let monotonicNanos = 5_000_000_000n;
-      const clock: Clock.Clock = {
-        currentTimeMillisUnsafe: () => wallMillis,
-        currentTimeMillis: Effect.sync(() => wallMillis),
-        currentTimeNanosUnsafe: () => monotonicNanos,
-        currentTimeNanos: Effect.sync(() => monotonicNanos),
-        sleep: () => Effect.void,
-      };
-      const overlayTimes: Array<number> = [];
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {
-        healthOverlay: (health, nowMillis) => {
-          overlayTimes.push(nowMillis);
-          return health;
-        },
-      }).pipe(Effect.provideService(Clock.Clock, clock));
-
-      wallMillis = 9_000;
-      monotonicNanos = 7_500_000_000n;
-      const refreshedHealth = yield* runtimeCore.refreshHealth.pipe(
-        Effect.provideService(Clock.Clock, clock),
-      );
-
-      expect(refreshedHealth.uptimeMs).toBe(2_500);
-      expect(overlayTimes).toStrictEqual([10_000, 9_000]);
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("records runtime core publish, engine mutation, and subscription fanout spans", () =>
-    Effect.gen(function* () {
-      const recording = makeRecordingTracer();
-      const observedSpans = yield* Effect.scoped(
-        Effect.gen(function* () {
-          const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {});
-          yield* Effect.addFinalizer(() => runtimeCore.close);
-          const subscription = yield* runtimeCore.liveClient.subscribe("orders", {
-            select: ["id", "price"],
-            orderBy: [{ field: "price", direction: "asc" }],
-            limit: 10,
-          });
-          yield* Effect.addFinalizer(() => subscription.close().pipe(Effect.orDie));
-
-          const eventsFiber = yield* subscription.events.pipe(
-            Stream.take(2),
-            Stream.runCollect,
-            Effect.forkChild,
-          );
-          yield* runtimeCore.client.publish("orders", order("a", 10));
-          const events = yield* Fiber.join(eventsFiber);
-          expect(events).toStrictEqual([
-            {
-              type: "snapshot",
-              topic: "orders",
-              queryId: "query-0",
-              version: 0,
-              keys: [],
-              rows: [],
-              totalRows: 0,
-            },
-            {
-              type: "delta",
-              topic: "orders",
-              queryId: "query-0",
-              fromVersion: 0,
-              toVersion: 1,
-              operations: [
-                {
-                  type: "insert",
-                  key: "a",
-                  row: {
-                    id: "a",
-                    price: 10,
-                  },
-                  index: 0,
-                },
-              ],
-              totalRows: 1,
-            },
-          ]);
-
-          return recording.spans;
-        }),
-      ).pipe(Effect.provideService(Tracer.Tracer, recording.tracer));
-
-      const spansByName = new Map(observedSpans.map((span) => [span.name, span]));
-      const clientPublish = spansByName.get("ViewServerRuntimeCore.client.publish");
-      const sourceMutationApply = spansByName.get("ViewServerRuntimeCore.sourceMutation.apply");
-      const publish = spansByName.get("ColumnLiveViewEngine.publish");
-      const topicStorePublish = spansByName.get("ColumnLiveViewEngine.topicStore.publish");
-      const mutationTransaction = spansByName.get(
-        "ColumnLiveViewEngine.topicStore.mutationTransaction",
-      );
-      const mutationBatch = spansByName.get("ColumnLiveViewEngine.topicStore.mutationBatch");
-      const notify = spansByName.get("ColumnLiveViewEngine.topicStore.notify");
-      const liveSubscriptionNotify = spansByName.get(
-        "ColumnLiveViewEngine.liveSubscription.notify",
-      );
-
-      expect({
-        clientPublish: {
-          name: clientPublish?.name,
-          parentSpanId: clientPublish?.parentSpanId,
-          traceId: clientPublish?.traceId,
-        },
-        liveSubscriptionNotify: {
-          attributes: liveSubscriptionNotify?.attributes,
-          name: liveSubscriptionNotify?.name,
-          parentName: liveSubscriptionNotify?.parentName,
-          parentSpanId: liveSubscriptionNotify?.parentSpanId,
-          traceId: liveSubscriptionNotify?.traceId,
-        },
-        mutationBatch: {
-          name: mutationBatch?.name,
-          parentName: mutationBatch?.parentName,
-          parentSpanId: mutationBatch?.parentSpanId,
-          traceId: mutationBatch?.traceId,
-        },
-        mutationTransaction: {
-          name: mutationTransaction?.name,
-          parentName: mutationTransaction?.parentName,
-          parentSpanId: mutationTransaction?.parentSpanId,
-          traceId: mutationTransaction?.traceId,
-        },
-        notify: {
-          name: notify?.name,
-          parentName: notify?.parentName,
-          parentSpanId: notify?.parentSpanId,
-          traceId: notify?.traceId,
-        },
-        publish: {
-          name: publish?.name,
-          parentName: publish?.parentName,
-          parentSpanId: publish?.parentSpanId,
-          traceId: publish?.traceId,
-        },
-        sourceMutationApply: {
-          name: sourceMutationApply?.name,
-          parentName: sourceMutationApply?.parentName,
-          parentSpanId: sourceMutationApply?.parentSpanId,
-          traceId: sourceMutationApply?.traceId,
-        },
-        topicStorePublish: {
-          name: topicStorePublish?.name,
-          parentName: topicStorePublish?.parentName,
-          parentSpanId: topicStorePublish?.parentSpanId,
-          traceId: topicStorePublish?.traceId,
-        },
-      }).toStrictEqual({
-        clientPublish: {
-          name: "ViewServerRuntimeCore.client.publish",
-          parentSpanId: null,
-          traceId: clientPublish?.traceId,
-        },
-        liveSubscriptionNotify: {
-          attributes: [
-            ["queryId", "query-0"],
-            ["topic", "orders"],
-          ],
-          name: "ColumnLiveViewEngine.liveSubscription.notify",
-          parentName: "ColumnLiveViewEngine.topicStore.notify",
-          parentSpanId: notify?.spanId,
-          traceId: clientPublish?.traceId,
-        },
-        mutationBatch: {
-          name: "ColumnLiveViewEngine.topicStore.mutationBatch",
-          parentName: "ColumnLiveViewEngine.topicStore.mutationTransaction",
-          parentSpanId: mutationTransaction?.spanId,
-          traceId: clientPublish?.traceId,
-        },
-        mutationTransaction: {
-          name: "ColumnLiveViewEngine.topicStore.mutationTransaction",
-          parentName: "ColumnLiveViewEngine.topicStore.publish",
-          parentSpanId: topicStorePublish?.spanId,
-          traceId: clientPublish?.traceId,
-        },
-        notify: {
-          name: "ColumnLiveViewEngine.topicStore.notify",
-          parentName: "ColumnLiveViewEngine.topicStore.mutationBatch",
-          parentSpanId: mutationBatch?.spanId,
-          traceId: clientPublish?.traceId,
-        },
-        publish: {
-          name: "ColumnLiveViewEngine.publish",
-          parentName: "ViewServerRuntimeCore.sourceMutation.apply",
-          parentSpanId: sourceMutationApply?.spanId,
-          traceId: clientPublish?.traceId,
-        },
-        sourceMutationApply: {
-          name: "ViewServerRuntimeCore.sourceMutation.apply",
-          parentName: "ViewServerRuntimeCore.client.publish",
-          parentSpanId: clientPublish?.spanId,
-          traceId: clientPublish?.traceId,
-        },
-        topicStorePublish: {
-          name: "ColumnLiveViewEngine.topicStore.publish",
-          parentName: "ColumnLiveViewEngine.publish",
-          parentSpanId: publish?.spanId,
-          traceId: clientPublish?.traceId,
-        },
-      });
-    }),
-  );
-
-  it.effect("runs the shared runtime core and live client", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {});
-      const subscription = yield* runtimeCore.liveClient.subscribe("orders", {
-        select: ["id", "price"],
-        orderBy: [{ field: "price", direction: "asc" }],
-        limit: 10,
-      });
-
-      const events = yield* subscription.events.pipe(Stream.take(1), Stream.runCollect);
-      expect(events[0]).toStrictEqual({
-        type: "snapshot",
-        topic: "orders",
-        queryId: "query-0",
-        version: 0,
-        keys: [],
-        rows: [],
-        totalRows: 0,
-      });
-
-      yield* runtimeCore.client.publishMany("orders", [order("b", 20), order("a", 10)]);
-      yield* runtimeCore.client.publish("orders", order("c", 30));
-      yield* runtimeCore.client.patch("orders", "c", { price: 5 });
-      yield* runtimeCore.client.delete("orders", "a");
-
-      const snapshot = yield* runtimeCore.client.snapshot("orders", {
-        select: ["id", "price"],
-        orderBy: [{ field: "price", direction: "asc" }],
-        limit: 10,
-      });
-      expect(snapshot.rows).toStrictEqual([
-        { id: "c", price: 5 },
-        { id: "b", price: 20 },
-      ]);
-
-      const health = yield* runtimeCore.client.health();
-      expect(health.engine.topics.orders.rowCount).toBe(2);
-      const refreshedHealth = yield* runtimeCore.refreshHealth;
-      expect(refreshedHealth.engine.topics.orders.rowCount).toBe(2);
-
-      yield* subscription.close();
-      yield* runtimeCore.client.reset();
-      const resetHealth = yield* runtimeCore.client.health();
-      expect(resetHealth.engine.topics.orders.rowCount).toBe(0);
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("exposes route-bypassing internals only through the internal factory", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCoreInternal(viewServer, {});
-      const subscription = yield* runtimeCore.internalLiveClient.subscribeInternal("orders", {
-        select: ["id"],
-      });
-      const events = yield* subscription.events.pipe(Stream.take(1), Stream.runCollect);
-
-      expect(events[0]).toStrictEqual({
-        type: "snapshot",
-        topic: "orders",
-        queryId: "query-0",
-        version: 0,
-        keys: [],
-        rows: [],
-        totalRows: 0,
-      });
-
-      yield* subscription.close();
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("supports the synchronous runtime core constructor", () =>
-    Effect.gen(function* () {
-      const runtimeCore = createViewServerRuntimeCore(viewServer, { subscriptionQueueCapacity: 1 });
-      expect("set" in runtimeCore.liveClient.health).toBe(false);
-      expect(runtimeCore.liveClient.health.value.status).toBe("ready");
-
-      yield* runtimeCore.client.publish("orders", order("a", 10));
-      const health = yield* runtimeCore.client.health();
-
-      expect(health.engine.topics.orders.rowCount).toBe(1);
-      expect(health.transport).toStrictEqual({
-        activeClients: 0,
-        activeStreams: 0,
-        activeSubscriptions: 0,
-        messagesPerSecond: 0,
-        bytesPerSecond: 0,
-        queuedMessages: 0,
-        queuedBytes: 0,
-        droppedClients: 0,
-        backpressureEvents: 0,
-        reconnects: 0,
-        lastError: null,
-      });
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("supports the synchronous runtime core constructor defaults", () =>
-    Effect.gen(function* () {
-      const runtimeCore = createViewServerRuntimeCore(viewServer);
-
-      yield* runtimeCore.client.publish("orders", order("a", 10));
-      const health = yield* runtimeCore.client.health();
-
-      expect(health.engine.topics.orders.rowCount).toBe(1);
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("rejects public leased topic queries before route validation", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(leasedViewServer, {});
-      const missingRouteQuery = {
-        where: {
-          region: { eq: "usa" },
-        },
-        select: ["id"],
-        limit: 1,
-      } as const;
-
-      const missingRouteEffect: Effect.Effect<unknown, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.client.snapshot,
-        runtimeCore.client,
-        ["orders", missingRouteQuery],
-      );
-      const missingRoute = yield* Effect.flip(missingRouteEffect);
-      expect(missingRoute).toStrictEqual(publicLeasedRuntimeAccessError);
-
-      const missingSubscribeRouteEffect: Effect.Effect<unknown, ViewServerRuntimeError> =
-        Reflect.apply(runtimeCore.liveClient.subscribe, runtimeCore.liveClient, [
-          "orders",
-          missingRouteQuery,
-        ]);
-      const missingSubscribeRoute = yield* Effect.flip(missingSubscribeRouteEffect);
-      expect(missingSubscribeRoute).toStrictEqual(publicLeasedRuntimeAccessError);
-
-      const nonEqRouteEffect: Effect.Effect<unknown, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.liveClient.subscribe,
-        runtimeCore.liveClient,
-        [
-          "orders",
-          {
-            where: {
-              region: { eq: "usa" },
-              status: { in: ["open"] },
-            },
-            select: ["id"],
-            limit: 1,
-          },
-        ],
-      );
-      const nonEqRoute = yield* Effect.flip(nonEqRouteEffect);
-      expect(nonEqRoute).toStrictEqual(publicLeasedRuntimeAccessError);
-
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("rejects direct leased topic runtime access through the public runtime core", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(leasedViewServer, {});
-      const leasedQuery = {
-        where: {
-          region: { eq: "usa" },
-          status: { eq: "open" },
-        },
-        select: ["id", "region", "status"],
-        limit: 1,
-      } as const;
-      const publishEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.client.publish,
-        runtimeCore.client,
-        ["orders", order("a", 10)],
-      );
-      const publishManyEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.client.publishMany,
-        runtimeCore.client,
-        ["orders", [order("b", 20)]],
-      );
-      const patchEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.client.patch,
-        runtimeCore.client,
-        ["orders", "a", { price: 20 }],
-      );
-      const deleteEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.client.delete,
-        runtimeCore.client,
-        ["orders", "a"],
-      );
-      const snapshotEffect: Effect.Effect<unknown, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.client.snapshot,
-        runtimeCore.client,
-        ["orders", leasedQuery],
-      );
-      const subscribeEffect: Effect.Effect<unknown, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.liveClient.subscribe,
-        runtimeCore.liveClient,
-        ["orders", leasedQuery],
-      );
-      const resetEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.client.reset,
-        runtimeCore.client,
-        [],
-      );
-      expect(yield* Effect.flip(publishEffect)).toStrictEqual(
-        publicSourceOwnedRuntimeMutationError,
-      );
-      expect(yield* Effect.flip(publishManyEffect)).toStrictEqual(
-        publicSourceOwnedRuntimeMutationError,
-      );
-      expect(yield* Effect.flip(patchEffect)).toStrictEqual(publicSourceOwnedRuntimeMutationError);
-      expect(yield* Effect.flip(deleteEffect)).toStrictEqual(publicSourceOwnedRuntimeMutationError);
-      expect(yield* Effect.flip(snapshotEffect)).toStrictEqual(publicLeasedRuntimeAccessError);
-      expect(yield* Effect.flip(subscribeEffect)).toStrictEqual(publicLeasedRuntimeAccessError);
-      expect(yield* Effect.flip(resetEffect)).toStrictEqual(publicSourceOwnedRuntimeResetError);
-
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("keeps source-owned materialized topics readable but blocks public mutations", () =>
-    Effect.gen(function* () {
-      const kafkaRuntimeCore = yield* makeViewServerRuntimeCoreInternal(kafkaOwnedViewServer, {});
-      const grpcRuntimeCore = yield* makeViewServerRuntimeCoreInternal(
-        materializedGrpcSourceViewServer,
-        {},
-      );
-
-      yield* kafkaRuntimeCore.internalClient.publish("orders", order("kafka", 10));
-      yield* grpcRuntimeCore.internalClient.publish("orders", order("grpc", 20));
-
-      const kafkaSnapshot = yield* kafkaRuntimeCore.publicClient.snapshot("orders", {
-        select: ["id", "price"],
-        limit: 10,
-      });
-      const grpcSnapshot = yield* grpcRuntimeCore.publicClient.snapshot("orders", {
-        select: ["id", "price"],
-        limit: 10,
-      });
-      const kafkaPublishEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        kafkaRuntimeCore.publicClient.publish,
-        kafkaRuntimeCore.publicClient,
-        ["orders", order("blocked-kafka", 30)],
-      );
-      const kafkaPublishManyEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        kafkaRuntimeCore.publicClient.publishMany,
-        kafkaRuntimeCore.publicClient,
-        ["orders", [order("blocked-kafka-many", 35)]],
-      );
-      const kafkaPatchEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        kafkaRuntimeCore.publicClient.patch,
-        kafkaRuntimeCore.publicClient,
-        ["orders", "kafka", { price: 35 }],
-      );
-      const kafkaDeleteEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        kafkaRuntimeCore.publicClient.delete,
-        kafkaRuntimeCore.publicClient,
-        ["orders", "kafka"],
-      );
-      const grpcPublishEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        grpcRuntimeCore.publicClient.publish,
-        grpcRuntimeCore.publicClient,
-        ["orders", order("blocked-grpc", 40)],
-      );
-      const grpcPublishManyEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        grpcRuntimeCore.publicClient.publishMany,
-        grpcRuntimeCore.publicClient,
-        ["orders", [order("blocked-grpc-many", 45)]],
-      );
-      const grpcPatchEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        grpcRuntimeCore.publicClient.patch,
-        grpcRuntimeCore.publicClient,
-        ["orders", "grpc", { price: 45 }],
-      );
-      const grpcDeleteEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        grpcRuntimeCore.publicClient.delete,
-        grpcRuntimeCore.publicClient,
-        ["orders", "grpc"],
-      );
-      const kafkaResetEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        kafkaRuntimeCore.publicClient.reset,
-        kafkaRuntimeCore.publicClient,
-        [],
-      );
-      const grpcResetEffect: Effect.Effect<void, ViewServerRuntimeError> = Reflect.apply(
-        grpcRuntimeCore.publicClient.reset,
-        grpcRuntimeCore.publicClient,
-        [],
-      );
-
-      expect(kafkaSnapshot).toStrictEqual({
-        rows: [{ id: "kafka", price: 10 }],
-        totalRows: 1,
-        version: 1,
-        status: "ready",
-        statusCode: "Ready",
-      });
-      expect(grpcSnapshot).toStrictEqual({
-        rows: [{ id: "grpc", price: 20 }],
-        totalRows: 1,
-        version: 1,
-        status: "ready",
-        statusCode: "Ready",
-      });
-      expect(yield* Effect.flip(kafkaPublishEffect)).toStrictEqual(
-        publicSourceOwnedRuntimeMutationError,
-      );
-      expect(yield* Effect.flip(kafkaPublishManyEffect)).toStrictEqual(
-        publicSourceOwnedRuntimeMutationError,
-      );
-      expect(yield* Effect.flip(kafkaPatchEffect)).toStrictEqual(
-        publicSourceOwnedRuntimeMutationError,
-      );
-      expect(yield* Effect.flip(kafkaDeleteEffect)).toStrictEqual(
-        publicSourceOwnedRuntimeMutationError,
-      );
-      expect(yield* Effect.flip(grpcPublishEffect)).toStrictEqual(
-        publicSourceOwnedRuntimeMutationError,
-      );
-      expect(yield* Effect.flip(grpcPublishManyEffect)).toStrictEqual(
-        publicSourceOwnedRuntimeMutationError,
-      );
-      expect(yield* Effect.flip(grpcPatchEffect)).toStrictEqual(
-        publicSourceOwnedRuntimeMutationError,
-      );
-      expect(yield* Effect.flip(grpcDeleteEffect)).toStrictEqual(
-        publicSourceOwnedRuntimeMutationError,
-      );
-      expect(yield* Effect.flip(kafkaResetEffect)).toStrictEqual(
-        publicSourceOwnedRuntimeResetError,
-      );
-      expect(yield* Effect.flip(grpcResetEffect)).toStrictEqual(publicSourceOwnedRuntimeResetError);
-
-      yield* kafkaRuntimeCore.close;
-      yield* grpcRuntimeCore.close;
-    }),
-  );
-
-  it.effect("rejects public grpcSource leased subscriptions before route validation", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(leasedGrpcSourceViewServer, {});
-      const leasedQuery = {
-        where: {
-          region: { eq: "usa" },
-          status: { eq: "open" },
-        },
-        select: ["id", "region", "status"],
-        limit: 1,
-      } as const;
-      const snapshotEffect: Effect.Effect<unknown, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.client.snapshot,
-        runtimeCore.client,
-        ["orders", leasedQuery],
-      );
-      const subscribeEffect: Effect.Effect<unknown, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.liveClient.subscribe,
-        runtimeCore.liveClient,
-        ["orders", leasedQuery],
-      );
-
-      expect(yield* Effect.flip(snapshotEffect)).toStrictEqual(publicLeasedRuntimeAccessError);
-      expect(yield* Effect.flip(subscribeEffect)).toStrictEqual(publicLeasedRuntimeAccessError);
-
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("allows internal runtime core access for leased gRPC manager internals", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCoreInternal(leasedViewServer, {});
-      yield* runtimeCore.internalClient.publish("orders", order("a", 10));
-      yield* runtimeCore.internalClient.publishManyWithStorageKeys("orders", [
-        {
-          storageKey: "orders/lease/row/public-b",
-          row: order("public-b", 20),
-        },
-      ]);
-
-      const snapshot = yield* runtimeCore.internalClient.snapshot("orders", {
-        where: {
-          customerId: { eq: "customer-a" },
-          region: { eq: "usa" },
-          status: { eq: "open" },
-        },
-        select: ["id", "region", "status"],
-        limit: 1,
-      });
-      const storageKeySnapshot = yield* runtimeCore.internalClient.snapshot("orders", {
-        where: {
-          customerId: { eq: "customer-public-b" },
-          region: { eq: "usa" },
-          status: { eq: "open" },
-        },
-        select: ["id", "region", "status"],
-        limit: 1,
-      });
-      const invalidRouteSnapshotEffect: Effect.Effect<unknown, ViewServerRuntimeError> =
-        Reflect.apply(runtimeCore.internalClient.snapshot, runtimeCore.internalClient, [
-          "orders",
-          {
-            where: {
-              region: { eq: "usa" },
-            },
-            select: ["id"],
-            limit: 1,
-          },
-        ]);
-      const invalidRouteSnapshot = yield* Effect.flip(invalidRouteSnapshotEffect);
-      const publicRuntimeSubscribe = yield* Effect.flip(
-        runtimeCore.liveClient.subscribeRuntime("orders", {
-          where: {
-            customerId: { eq: "customer-a" },
-            region: { eq: "usa" },
-            status: { eq: "open" },
-          },
-          select: ["id"],
-          limit: 1,
-        }),
-      );
-      const subscription = yield* runtimeCore.internalLiveClient.subscribeInternal("orders", {
-        where: {
-          customerId: { eq: "customer-a" },
-          region: { eq: "usa" },
-          status: { eq: "open" },
-        },
-        select: ["id"],
-        limit: 1,
-      });
-      const events = yield* subscription.events.pipe(Stream.take(1), Stream.runCollect);
-
-      expect(snapshot).toStrictEqual({
-        rows: [
-          {
-            id: "a",
-            region: "usa",
-            status: "open",
-          },
-        ],
-        totalRows: 1,
-        version: 2,
-        status: "ready",
-        statusCode: "Ready",
-      });
-      expect(storageKeySnapshot).toStrictEqual({
-        rows: [
-          {
-            id: "public-b",
-            region: "usa",
-            status: "open",
-          },
-        ],
-        totalRows: 1,
-        version: 2,
-        status: "ready",
-        statusCode: "Ready",
-      });
-      expect(invalidRouteSnapshot).toStrictEqual({
-        _tag: "ViewServerRuntimeError",
-        code: "InvalidQuery",
-        topic: "orders",
-        message: "Leased topic orders route field status must use an exact eq filter.",
-      });
-      expect(publicRuntimeSubscribe).toStrictEqual(publicLeasedRuntimeAccessError);
-      expect(events[0]).toStrictEqual({
-        type: "snapshot",
-        topic: "orders",
-        queryId: "query-0",
-        version: 2,
-        keys: ["a"],
-        rows: [{ id: "a" }],
-        totalRows: 1,
-      });
-
-      yield* subscription.close();
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("allows internal decoded row publishing for source-owned runtime internals", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCoreInternal(viewServer, {});
-      yield* runtimeCore.internalClient.publishManyDecodedRows("orders", [order("decoded", 30)]);
-      yield* runtimeCore.internalClient.publishManyDecodedRowsWithStorageKeys("orders", [
-        {
-          storageKey: "orders/source/row/storage-decoded",
-          row: order("public-decoded", 40),
-        },
-      ]);
-
-      const decodedSnapshot = yield* runtimeCore.internalClient.snapshot("orders", {
-        where: {
-          customerId: { eq: "customer-decoded" },
-        },
-        select: ["id", "price"],
-        limit: 1,
-      });
-      const storageKeySnapshot = yield* runtimeCore.internalClient.snapshot("orders", {
-        where: {
-          customerId: { eq: "customer-public-decoded" },
-        },
-        select: ["id", "price"],
-        limit: 1,
-      });
-
-      expect(decodedSnapshot).toStrictEqual({
-        rows: [{ id: "decoded", price: 30 }],
-        totalRows: 1,
-        version: 2,
-        status: "ready",
-        statusCode: "Ready",
-      });
-      expect(storageKeySnapshot).toStrictEqual({
-        rows: [{ id: "public-decoded", price: 40 }],
-        totalRows: 1,
-        version: 2,
-        status: "ready",
-        statusCode: "Ready",
-      });
-
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("allows internal decoded field patching for source-owned runtime internals", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCoreInternal(viewServer, {});
-      yield* runtimeCore.internalClient.publishManyDecodedRows("orders", [order("decoded", 30)]);
-      yield* runtimeCore.internalClient.patchDecodedFields("orders", "decoded", {
-        price: 31,
-        status: "closed",
-      });
-      const snapshot = yield* runtimeCore.internalClient.snapshot("orders", {
-        select: ["id", "price", "status"],
-        limit: 1,
-      });
-
-      expect(snapshot).toStrictEqual({
-        rows: [{ id: "decoded", price: 31, status: "closed" }],
-        totalRows: 1,
-        version: 2,
-        status: "ready",
-        statusCode: "Ready",
-      });
-
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("rejects public leased subscriptions when effects execute", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(leasedViewServer, {});
-      const delayedSubscribeQuery = {
-        where: {
-          region: { eq: "usa" },
-          status: { eq: "open" },
-        },
-        select: ["id"],
-        limit: 1,
-      } satisfies {
-        readonly where: {
-          readonly region: {
-            readonly eq: "usa";
-          };
-          readonly status: {
-            readonly eq: "open";
-          };
-        };
-        readonly select: readonly ["id"];
-        readonly limit: 1;
-      };
-      const subscribeEffect: Effect.Effect<unknown, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.liveClient.subscribe,
-        runtimeCore.liveClient,
-        ["orders", delayedSubscribeQuery],
-      );
-      expect(Reflect.deleteProperty(delayedSubscribeQuery.where.status, "eq")).toBe(true);
-      Object.defineProperty(delayedSubscribeQuery.where.status, "in", {
-        enumerable: true,
-        value: ["open"],
-      });
-
-      const subscribeRouteError = yield* Effect.flip(subscribeEffect);
-      expect(subscribeRouteError).toStrictEqual(publicLeasedRuntimeAccessError);
-
-      const delayedRuntimeQuery = {
-        where: {
-          region: { eq: "usa" },
-          status: { eq: "open" },
-        },
-        select: ["id"],
-        limit: 1,
-      } satisfies {
-        readonly where: {
-          readonly region: {
-            readonly eq: "usa";
-          };
-          readonly status: {
-            readonly eq: "open";
-          };
-        };
-        readonly select: readonly ["id"];
-        readonly limit: 1;
-      };
-      const subscribeRuntimeEffect: Effect.Effect<unknown, ViewServerRuntimeError> = Reflect.apply(
-        runtimeCore.liveClient.subscribe,
-        runtimeCore.liveClient,
-        ["orders", delayedRuntimeQuery],
-      );
-      expect(Reflect.deleteProperty(delayedRuntimeQuery.where.status, "eq")).toBe(true);
-      Object.defineProperty(delayedRuntimeQuery.where.status, "in", {
-        enumerable: true,
-        value: ["open"],
-      });
-
-      const subscribeRuntimeRouteError = yield* Effect.flip(subscribeRuntimeEffect);
-      expect(subscribeRuntimeRouteError).toStrictEqual(publicLeasedRuntimeAccessError);
-
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("forwards grouped admission limits to the engine", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {
-        groupedIncrementalAdmissionLimits: {
-          maxGroups: 1,
-        },
-      });
-      yield* runtimeCore.client.publishMany("orders", [order("a", 10), order("b", 20)]);
-      const subscription = yield* runtimeCore.liveClient.subscribe("orders", {
-        groupBy: ["price"],
-        aggregates: {
-          rowCount: { aggFunc: "count" },
-        },
-        orderBy: [{ field: "price", direction: "asc" }],
-        limit: 10,
-      });
-
-      const health = yield* runtimeCore.client.health();
-      expect(health.engine.topics.orders.activeFallbackGroupedViews).toBe(1);
-      expect(health.engine.topics.orders.activeIncrementalGroupedViews).toBe(0);
-
-      yield* subscription.close();
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("subscribes through the runtime live-client entrypoint", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCoreInternal(viewServer, {});
-      yield* runtimeCore.internalClient.publish("orders", order("a", 10));
-
-      const subscription = yield* runtimeCore.liveClient.subscribeRuntime("orders", {
-        select: ["id", "price"],
-      });
-      const events = yield* subscription.events.pipe(Stream.take(1), Stream.runCollect);
-
-      expect(events[0]).toStrictEqual({
-        type: "snapshot",
-        topic: "orders",
-        queryId: "query-0",
-        version: 1,
-        keys: ["a"],
-        rows: [{ id: "a", price: 10 }],
-        totalRows: 1,
-      });
-
-      yield* subscription.close();
-      const health = yield* runtimeCore.internalClient.health();
-      expect(health.engine.topics.orders.activeSubscriptions).toBe(0);
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("forwards producer terminal observation through the internal live client", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCoreInternal(viewServer, {
-        subscriptionQueueCapacity: 1,
-      });
-      const phases: Array<string> = [];
-      const subscription = yield* runtimeCore.internalLiveClient.subscribeObservedInternal(
-        "orders",
-        {
-          select: ["id"],
-          limit: 10,
-        },
-        {
-          onQueryRegistered: (queryId) =>
-            Effect.sync(() => {
-              phases.push(`registered:${queryId}`);
-            }),
-          onTerminalOccurrence: () =>
-            Effect.sync(() => {
-              phases.push("occurrence");
-            }),
-          onTerminalReady: () =>
-            Effect.sync(() => {
-              phases.push("ready");
-            }),
-        },
-      );
-
-      yield* runtimeCore.internalClient.publish("orders", order("observed", 10));
-      const events = yield* subscription.events.pipe(Stream.runCollect);
-      expect(phases).toStrictEqual(["registered:query-0", "occurrence", "ready"]);
-      expect(Array.from(events)).toStrictEqual([
-        {
-          type: "status",
-          topic: "orders",
-          queryId: "query-0",
-          status: "closed",
-          code: "BackpressureExceeded",
-          message: "Subscription closed because its event queue exceeded capacity.",
-        },
-      ]);
-
-      const runtimeQueryIds: Array<string> = [];
-      const runtimeSubscription =
-        yield* runtimeCore.internalLiveClient.subscribeRuntimeObservedInternal(
-          "orders",
-          { select: ["id"] },
-          {
-            onQueryRegistered: (queryId) =>
-              Effect.sync(() => {
-                runtimeQueryIds.push(queryId);
-              }),
-            onTerminalOccurrence: () => Effect.void,
-            onTerminalReady: () => Effect.void,
-          },
-        );
-      expect(runtimeQueryIds).toStrictEqual(["query-1"]);
-      yield* runtimeSubscription.close();
-      yield* subscription.close();
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("releases observed subscriptions when initial health refresh fails", () =>
-    Effect.gen(function* () {
-      const engine = yield* createColumnLiveViewEngineInternal({ topics: viewServer.topics });
-      const health = AtomRef.make(healthFromEngine(yield* engine.health()));
-      const liveClient = yield* makeRuntimeCoreLiveClient(
-        viewServer,
-        engine,
-        health,
-        Effect.fail(refreshFailed),
-      );
-      const observer = {
-        onQueryRegistered: () => Effect.void,
-        onTerminalOccurrence: () => Effect.void,
-        onTerminalReady: () => Effect.void,
-      };
-
-      const failedTyped = yield* liveClient
-        .subscribeObservedInternal(
-          "orders",
-          {
-            select: ["id"],
-            limit: 1,
-          },
-          observer,
-        )
-        .pipe(Effect.flip);
-      const afterTypedFailure = yield* engine.health();
-      const failedRuntime = yield* liveClient
-        .subscribeRuntimeObservedInternal(
-          "orders",
-          {
-            select: ["id"],
-            limit: 1,
-          },
-          observer,
-        )
-        .pipe(Effect.flip);
-      const afterRuntimeFailure = yield* engine.health();
-
-      expect({
-        failedTyped,
-        typedSubscriptions: afterTypedFailure.activeSubscriptions,
-        failedRuntime,
-        runtimeSubscriptions: afterRuntimeFailure.activeSubscriptions,
-      }).toStrictEqual({
-        failedTyped: refreshFailed,
-        typedSubscriptions: 0,
-        failedRuntime: refreshFailed,
-        runtimeSubscriptions: 0,
-      });
-      yield* liveClient.close;
-    }),
-  );
-
-  it.effect("releases acquired live subscriptions when initial health refresh fails", () =>
-    Effect.gen(function* () {
-      const engine = yield* createColumnLiveViewEngineInternal({ topics: viewServer.topics });
-      const health = AtomRef.make(healthFromEngine(yield* engine.health()));
-      const liveClient = yield* makeRuntimeCoreLiveClient(
-        viewServer,
-        engine,
-        health,
-        Effect.fail(refreshFailed),
-      );
-
-      const failedRaw = yield* Effect.flip(
-        liveClient.subscribe("orders", {
-          select: ["id"],
-          limit: 1,
-        }),
-      );
-      const afterRawFailure = yield* engine.health();
-      expect(failedRaw).toStrictEqual(refreshFailed);
-      expect(afterRawFailure.activeSubscriptions).toBe(0);
-
-      const failedRuntime = yield* Effect.flip(
-        liveClient.subscribeRuntime("orders", {
-          select: ["id"],
-          limit: 1,
-        }),
-      );
-      const afterRuntimeFailure = yield* engine.health();
-      expect(failedRuntime).toStrictEqual(refreshFailed);
-      expect(afterRuntimeFailure.activeSubscriptions).toBe(0);
-
-      yield* liveClient.close;
-    }),
-  );
-
-  it.effect("releases pushed health subscriptions when initial health refresh fails", () =>
-    Effect.gen(function* () {
-      const engine = yield* createColumnLiveViewEngineInternal({ topics: viewServer.topics });
-      const health = AtomRef.make(healthFromEngine(yield* engine.health()));
-      const liveClient = yield* makeRuntimeCoreLiveClient(
-        viewServer,
-        engine,
-        health,
-        Effect.fail(refreshFailed),
-      );
-
-      const failedSummary = yield* Effect.flip(liveClient.subscribeHealthSummary());
-      const failedDetail = yield* Effect.flip(liveClient.subscribeHealth());
-
-      expect(failedSummary).toStrictEqual(refreshFailed);
-      expect(failedDetail).toStrictEqual(refreshFailed);
-      yield* liveClient.close.pipe(Effect.timeout("1 second"));
-    }),
-  );
-
-  it.effect("refreshes health after close", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {
-        healthRefreshCadence: "0 millis",
-      });
-      yield* runtimeCore.client.publish("orders", order("a", 10));
-
-      const ready = yield* runtimeCore.client.health();
-      expect(ready.status).toBe("ready");
-
-      yield* runtimeCore.liveClient.close;
-
-      const closed = yield* runtimeCore.client.health();
-      expect(closed.status).toBe("stopping");
-      expect(closed.engine.topics.orders.activeSubscriptions).toBe(0);
-    }),
-  );
-
-  it.effect("live client close owns pending runtime health refresh cleanup", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {
-        healthRefreshCadence: "1 minute",
-      });
-
-      yield* runtimeCore.client.publish("orders", order("a", 10));
-      yield* runtimeCore.liveClient.close;
-
-      const closed = yield* runtimeCore.client.health();
-      expect(closed.status).toBe("stopping");
-      expect(closed.engine.topics.orders.rowCount).toBe(1);
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("pushes summary and detailed health snapshots", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {
-        healthRefreshCadence: "0 millis",
-      });
-      const summary = yield* runtimeCore.liveClient.subscribeHealthSummary();
-      const detail = yield* runtimeCore.liveClient.subscribeHealth();
-
-      const summaryFiber = yield* summary.events.pipe(
-        Stream.take(2),
-        Stream.runCollect,
-        Effect.forkChild,
-      );
-      const detailFiber = yield* detail.events.pipe(
-        Stream.take(2),
-        Stream.runCollect,
-        Effect.forkChild,
-      );
-
-      yield* runtimeCore.client.publish("orders", order("a", 10));
-
-      const summaryEvents = yield* Fiber.join(summaryFiber);
-      const detailEvents = yield* Fiber.join(detailFiber);
-      expect(Array.from(summaryEvents)).toStrictEqual([
-        {
-          type: "snapshot",
-          topic: "__view_server_health_summary",
-          queryId: "health-summary",
-          version: 0,
-          keys: ["summary"],
-          rows: [
-            {
-              id: "summary",
-              status: "ready",
-              runtimeStatus: "ready",
-              connectionStatus: "connected",
-              unhealthyTopics: [],
-              updatedAtNanos: expect.anything(),
-              maxKafkaLag: null,
-            },
-          ],
-          totalRows: 1,
-        },
-        {
-          type: "snapshot",
-          topic: "__view_server_health_summary",
-          queryId: "health-summary",
-          version: 1,
-          keys: ["summary"],
-          rows: [
-            {
-              id: "summary",
-              status: "ready",
-              runtimeStatus: "ready",
-              connectionStatus: "connected",
-              unhealthyTopics: [],
-              updatedAtNanos: expect.anything(),
-              maxKafkaLag: null,
-            },
-          ],
-          totalRows: 1,
-        },
-      ]);
-      expect(Array.from(detailEvents)).toStrictEqual([
-        {
-          type: "snapshot",
-          topic: "__view_server_health",
-          queryId: "health",
-          version: 0,
-          keys: ["orders"],
-          rows: [
-            {
-              id: "orders",
-              status: "ready",
-              rowCount: 0,
-              liveRowCount: 0,
-              deletedRowCount: 0,
-              version: 0,
-              lastMutationAt: null,
-              mutationsPerSecond: 0,
-              rowsPerSecond: 0,
-              pendingMutationBatches: 0,
-              activeFallbackGroupedViews: 0,
-              activeIncrementalGroupedViews: 0,
-              activeViews: 0,
-              groupedFullEvaluationCount: 0,
-              groupedPatchedEvaluationCount: 0,
-              activeSubscriptions: 0,
-              queuedEvents: 0,
-              maxQueueDepth: 0,
-              backpressureEvents: 0,
-              memoryBytes: 0,
-              tombstoneCount: 0,
-              compactionPending: false,
-              kafkaLag: null,
-              updatedAtNanos: expect.anything(),
-            },
-          ],
-          totalRows: 1,
-        },
-        {
-          type: "snapshot",
-          topic: "__view_server_health",
-          queryId: "health",
-          version: 1,
-          keys: ["orders"],
-          rows: [
-            {
-              id: "orders",
-              status: "ready",
-              rowCount: 1,
-              liveRowCount: 1,
-              deletedRowCount: 0,
-              version: 1,
-              lastMutationAt: expect.anything(),
-              mutationsPerSecond: 1,
-              rowsPerSecond: 1,
-              pendingMutationBatches: 0,
-              activeFallbackGroupedViews: 0,
-              activeIncrementalGroupedViews: 0,
-              activeViews: 0,
-              groupedFullEvaluationCount: 0,
-              groupedPatchedEvaluationCount: 0,
-              activeSubscriptions: 0,
-              queuedEvents: 0,
-              maxQueueDepth: 0,
-              backpressureEvents: 0,
-              memoryBytes: 0,
-              tombstoneCount: 0,
-              compactionPending: false,
-              kafkaLag: null,
-              updatedAtNanos: expect.anything(),
-            },
-          ],
-          totalRows: 1,
-        },
-      ]);
-
-      yield* summary.close();
-      yield* summary.close();
-      yield* detail.close();
-      yield* detail.close();
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("applies health overlays to pushed health subscriptions", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {
-        healthOverlay: (health) => ({
-          ...health,
-          status: "degraded",
-          kafka: {
-            startFrom: {
-              consumerGroupId: "view-server-test",
-              fallbackMode: "earliest",
-              mode: "committed",
-            },
-            regions: {
-              local: {
-                status: "connected",
-                brokers: "localhost:9092",
-                lastConnectedAt: 1_000,
-                lastError: null,
-              },
-            },
-            topics: {
-              sourceOrders: {
-                status: "stalled",
-                sourceTopic: "orders-source",
-                viewServerTopic: "orders",
-                regions: {
-                  local: {
-                    connected: true,
-                    assignedPartitions: 1,
-                    messagesPerSecond: 0,
-                    bytesPerSecond: 0,
-                    decodedMessagesPerSecond: 0,
-                    decodeFailuresPerSecond: 0,
-                    mappingFailuresPerSecond: 0,
-                    publishFailuresPerSecond: 0,
-                    commitFailuresPerSecond: 0,
-                    processingFailuresPerSecond: 0,
-                    lastMessageAt: null,
-                    lastCommitAt: null,
-                    consumerLagMessages: 7n,
-                    lagSampledAt: null,
-                    committedOffset: "3",
-                    lastError: "lagging",
-                  },
-                },
-              },
-            },
-          },
-        }),
-      });
-      const summary = yield* runtimeCore.liveClient.subscribeHealthSummary();
-      const detail = yield* runtimeCore.liveClient.subscribeHealth();
-
-      const summaryEvents = yield* summary.events.pipe(Stream.take(1), Stream.runCollect);
-      const detailEvents = yield* detail.events.pipe(Stream.take(1), Stream.runCollect);
-
-      expect(Array.from(summaryEvents)).toStrictEqual([
-        {
-          type: "snapshot",
-          topic: "__view_server_health_summary",
-          queryId: "health-summary",
-          version: 0,
-          keys: ["summary"],
-          rows: [
-            {
-              id: "summary",
-              status: "degraded",
-              runtimeStatus: "degraded",
-              connectionStatus: "connected",
-              unhealthyTopics: ["orders"],
-              updatedAtNanos: expect.anything(),
-              maxKafkaLag: 7n,
-            },
-          ],
-          totalRows: 1,
-        },
-      ]);
-      expect(Array.from(detailEvents)).toStrictEqual([
-        {
-          type: "snapshot",
-          topic: "__view_server_health",
-          queryId: "health",
-          version: 0,
-          keys: ["orders"],
-          rows: [
-            {
-              id: "orders",
-              status: "degraded",
-              rowCount: 0,
-              liveRowCount: 0,
-              deletedRowCount: 0,
-              version: 0,
-              lastMutationAt: null,
-              mutationsPerSecond: 0,
-              rowsPerSecond: 0,
-              pendingMutationBatches: 0,
-              activeFallbackGroupedViews: 0,
-              activeIncrementalGroupedViews: 0,
-              activeViews: 0,
-              groupedFullEvaluationCount: 0,
-              groupedPatchedEvaluationCount: 0,
-              activeSubscriptions: 0,
-              queuedEvents: 0,
-              maxQueueDepth: 0,
-              backpressureEvents: 0,
-              memoryBytes: 0,
-              tombstoneCount: 0,
-              compactionPending: false,
-              kafkaLag: 7n,
-              updatedAtNanos: expect.anything(),
-            },
-          ],
-          totalRows: 1,
-        },
-      ]);
-
-      yield* summary.close();
-      yield* detail.close();
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("closes slow pushed health summary subscriptions with typed backpressure", () =>
-    Effect.gen(function* () {
-      const engine = yield* createColumnLiveViewEngineInternal({ topics: viewServer.topics });
-      const health = AtomRef.make(healthFromEngine(yield* engine.health()));
-      const liveClient = yield* makeRuntimeCoreLiveClient(
-        viewServer,
-        engine,
-        health,
-        Effect.sync(() => health.value),
-      );
-      const summary = yield* liveClient.subscribeHealthSummary();
-
-      yield* Effect.forEach(
-        Array.from({ length: 64 }, (_, index) => index),
-        (index) =>
-          Effect.sync(() => {
-            health.update(() => healthFromEngine(engineHealth("ready", index + 1)));
-          }).pipe(Effect.andThen(Effect.yieldNow)),
-      );
-      const events = yield* summary.events.pipe(Stream.runCollect, Effect.timeout("1 second"));
-
-      expect(Array.from(events)).toStrictEqual([
-        {
-          type: "status",
-          topic: "__view_server_health_summary",
-          queryId: "health-summary",
-          status: "closed",
-          code: "BackpressureExceeded",
-          message:
-            "Runtime health subscription closed because its event queue exceeded capacity with 64 queued event(s).",
-        },
-      ]);
-
-      yield* summary.close();
-      yield* liveClient.close;
-    }),
-  );
-
-  it.effect(
-    "keeps detailed health subscription acquisition stable when health changes during refresh",
-    () =>
-      Effect.gen(function* () {
-        const engine = yield* createColumnLiveViewEngineInternal({ topics: viewServer.topics });
-        const health = AtomRef.make(healthFromEngine(yield* engine.health()));
-        const refreshStarted = yield* Deferred.make<void>();
-        const releaseRefresh = yield* Deferred.make<void>();
-        const liveClient = yield* makeRuntimeCoreLiveClient(
-          viewServer,
-          engine,
-          health,
-          Effect.gen(function* () {
-            yield* Deferred.succeed(refreshStarted, undefined);
-            yield* Deferred.await(releaseRefresh);
-            return health.value;
-          }),
-        );
-
-        const detailFiber = yield* liveClient
-          .subscribeHealth()
-          .pipe(Effect.forkChild({ startImmediately: true }));
-        yield* Deferred.await(refreshStarted).pipe(Effect.timeout("1 second"));
-        yield* Effect.forEach(
-          Array.from({ length: 128 }, (_, index) => index),
-          (index) =>
-            Effect.sync(() => {
-              health.update(() => healthFromEngine(engineHealth("ready", index + 1)));
-            }),
-        );
-        yield* Deferred.succeed(releaseRefresh, undefined);
-
-        const detail = yield* Fiber.join(detailFiber).pipe(Effect.timeout("1 second"));
-        const events = yield* detail.events.pipe(Stream.take(2), Stream.runCollect);
-
-        expect(Array.from(events)).toStrictEqual([
-          {
-            type: "snapshot",
-            topic: "__view_server_health",
-            queryId: "health",
-            version: 1,
-            keys: ["orders"],
-            rows: [
-              {
-                id: "orders",
-                status: "ready",
-                rowCount: 128,
-                liveRowCount: 128,
-                deletedRowCount: 0,
-                version: 1,
-                lastMutationAt: null,
-                mutationsPerSecond: 0,
-                rowsPerSecond: 0,
-                pendingMutationBatches: 0,
-                activeFallbackGroupedViews: 0,
-                activeIncrementalGroupedViews: 0,
-                activeViews: 0,
-                groupedFullEvaluationCount: 0,
-                groupedPatchedEvaluationCount: 0,
-                activeSubscriptions: 0,
-                queuedEvents: 0,
-                maxQueueDepth: 0,
-                backpressureEvents: 0,
-                memoryBytes: 0,
-                tombstoneCount: 0,
-                compactionPending: false,
-                kafkaLag: null,
-                updatedAtNanos: expect.anything(),
-              },
-            ],
-            totalRows: 1,
-          },
-          {
-            type: "snapshot",
-            topic: "__view_server_health",
-            queryId: "health",
-            version: 1,
-            keys: ["orders"],
-            rows: [
-              {
-                id: "orders",
-                status: "ready",
-                rowCount: 128,
-                liveRowCount: 128,
-                deletedRowCount: 0,
-                version: 1,
-                lastMutationAt: null,
-                mutationsPerSecond: 0,
-                rowsPerSecond: 0,
-                pendingMutationBatches: 0,
-                activeFallbackGroupedViews: 0,
-                activeIncrementalGroupedViews: 0,
-                activeViews: 0,
-                groupedFullEvaluationCount: 0,
-                groupedPatchedEvaluationCount: 0,
-                activeSubscriptions: 0,
-                queuedEvents: 0,
-                maxQueueDepth: 0,
-                backpressureEvents: 0,
-                memoryBytes: 0,
-                tombstoneCount: 0,
-                compactionPending: false,
-                kafkaLag: null,
-                updatedAtNanos: expect.anything(),
-              },
-            ],
-            totalRows: 1,
-          },
-        ]);
-
-        yield* detail.close();
-        yield* liveClient.close;
-      }),
-  );
-
-  it.effect(
-    "rejects pending pushed health subscriptions when live client closes during refresh",
-    () =>
-      Effect.gen(function* () {
-        const engine = yield* createColumnLiveViewEngineInternal({ topics: viewServer.topics });
-        const health = AtomRef.make(healthFromEngine(yield* engine.health()));
-        const refreshStarted = yield* Deferred.make<void>();
-        const releaseRefresh = yield* Deferred.make<void>();
-        const liveClient = yield* makeRuntimeCoreLiveClient(
-          viewServer,
-          engine,
-          health,
-          Effect.gen(function* () {
-            yield* Deferred.succeed(refreshStarted, undefined);
-            yield* Deferred.await(releaseRefresh);
-            return health.value;
-          }),
-        );
-
-        const summaryFiber = yield* liveClient
-          .subscribeHealthSummary()
-          .pipe(Effect.flip, Effect.forkChild({ startImmediately: true }));
-        yield* Deferred.await(refreshStarted).pipe(Effect.timeout("1 second"));
-        const closeFiber = yield* liveClient.close.pipe(
-          Effect.forkChild({ startImmediately: true }),
-        );
-        yield* Deferred.succeed(releaseRefresh, undefined);
-
-        const closedSummary = yield* Fiber.join(summaryFiber).pipe(Effect.timeout("1 second"));
-        yield* Fiber.join(closeFiber).pipe(Effect.timeout("1 second"));
-
-        expect(closedSummary).toStrictEqual({
-          _tag: "ViewServerRuntimeError",
-          code: "RuntimeUnavailable",
-          message: "Runtime Core is closed.",
-        });
-      }),
-  );
-
-  it.effect("closes slow runtime health summary subscriptions from real health refreshes", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {
-        healthRefreshCadence: "1 minute",
-      });
-      const summary = yield* runtimeCore.liveClient.subscribeHealthSummary();
-
-      yield* Effect.forEach(
-        Array.from({ length: 64 }, (_, index) => index),
-        (index) =>
-          runtimeCore.client
-            .publish("orders", order(`slow-health-${index}`, index))
-            .pipe(Effect.andThen(runtimeCore.client.health()), Effect.andThen(Effect.yieldNow)),
-      );
-      const events = yield* summary.events.pipe(Stream.runCollect, Effect.timeout("1 second"));
-
-      expect(Array.from(events)).toStrictEqual([
-        {
-          type: "status",
-          topic: "__view_server_health_summary",
-          queryId: "health-summary",
-          status: "closed",
-          code: "BackpressureExceeded",
-          message:
-            "Runtime health subscription closed because its event queue exceeded capacity with 64 queued event(s).",
-        },
-      ]);
-
-      yield* summary.close();
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("closes slow pushed detailed health subscriptions with typed backpressure", () =>
-    Effect.gen(function* () {
-      const engine = yield* createColumnLiveViewEngineInternal({ topics: viewServer.topics });
-      const health = AtomRef.make(healthFromEngine(yield* engine.health()));
-      const liveClient = yield* makeRuntimeCoreLiveClient(
-        viewServer,
-        engine,
-        health,
-        Effect.sync(() => health.value),
-      );
-      const detail = yield* liveClient.subscribeHealth();
-
-      yield* Effect.forEach(
-        Array.from({ length: 64 }, (_, index) => index),
-        (index) =>
-          Effect.sync(() => {
-            health.update(() => healthFromEngine(engineHealth("ready", index + 1)));
-          }).pipe(Effect.andThen(Effect.yieldNow)),
-      );
-      const events = yield* detail.events.pipe(Stream.runCollect, Effect.timeout("1 second"));
-
-      expect(Array.from(events)).toStrictEqual([
-        {
-          type: "status",
-          topic: "__view_server_health",
-          queryId: "health",
-          status: "closed",
-          code: "BackpressureExceeded",
-          message:
-            "Runtime health subscription closed because its event queue exceeded capacity with 64 queued event(s).",
-        },
-      ]);
-
-      yield* detail.close();
-      yield* liveClient.close;
-    }),
-  );
-
-  it.effect("closes active pushed health subscriptions when the live client closes", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {});
-      const summary = yield* runtimeCore.liveClient.subscribeHealthSummary();
-      const detail = yield* runtimeCore.liveClient.subscribeHealth();
-      const summaryFiber = yield* summary.events.pipe(Stream.runDrain, Effect.forkChild);
-      const detailFiber = yield* detail.events.pipe(Stream.runDrain, Effect.forkChild);
-
-      yield* Effect.yieldNow;
-      yield* runtimeCore.close;
-      yield* Fiber.join(summaryFiber);
-      yield* Fiber.join(detailFiber);
-
-      const closed = yield* runtimeCore.client.health();
-      expect(closed.status).toBe("stopping");
-    }),
-  );
-
-  it.effect("rejects pushed health subscriptions after live client close", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {});
-
-      yield* runtimeCore.liveClient.close;
-      const closedSummary = yield* Effect.flip(runtimeCore.liveClient.subscribeHealthSummary());
-      const closedDetail = yield* Effect.flip(runtimeCore.liveClient.subscribeHealth());
-
-      expect(closedSummary).toStrictEqual({
-        _tag: "ViewServerRuntimeError",
-        code: "RuntimeUnavailable",
-        message: "Runtime Core is closed.",
-      });
-      expect(closedDetail).toStrictEqual({
-        _tag: "ViewServerRuntimeError",
-        code: "RuntimeUnavailable",
-        message: "Runtime Core is closed.",
-      });
-      yield* runtimeCore.close;
-    }),
-  );
-
-  it.effect("keeps pushed health subscriptions alive after the acquisition fiber completes", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {});
-      const acquisitionFiber = yield* runtimeCore.liveClient
-        .subscribeHealthSummary()
-        .pipe(Effect.forkChild);
-      const summary = yield* Fiber.join(acquisitionFiber);
-      const eventsFiber = yield* summary.events.pipe(
-        Stream.take(2),
-        Stream.runCollect,
-        Effect.forkChild,
-      );
-
-      yield* runtimeCore.client.publish("orders", order("a", 10));
-      yield* runtimeCore.client.health();
-      const events = yield* Fiber.join(eventsFiber).pipe(
-        Effect.timeout("1 second"),
-        Effect.ensuring(summary.close().pipe(Effect.orDie, Effect.andThen(runtimeCore.close))),
-      );
-
-      expect(Array.from(events)).toStrictEqual([
-        {
-          type: "snapshot",
-          topic: "__view_server_health_summary",
-          queryId: "health-summary",
-          version: 0,
-          keys: ["summary"],
-          rows: [
-            {
-              id: "summary",
-              status: "ready",
-              runtimeStatus: "ready",
-              connectionStatus: "connected",
-              unhealthyTopics: [],
-              updatedAtNanos: expect.anything(),
-              maxKafkaLag: null,
-            },
-          ],
-          totalRows: 1,
-        },
-        {
-          type: "snapshot",
-          topic: "__view_server_health_summary",
-          queryId: "health-summary",
-          version: 1,
-          keys: ["summary"],
-          rows: [
-            {
-              id: "summary",
-              status: "ready",
-              runtimeStatus: "ready",
-              connectionStatus: "connected",
-              unhealthyTopics: [],
-              updatedAtNanos: expect.anything(),
-              maxKafkaLag: null,
-            },
-          ],
-          totalRows: 1,
-        },
-      ]);
-    }),
-  );
-
-  it.effect("does not let stale detached health refreshes overwrite stopping health", () =>
-    Effect.gen(function* () {
-      const firstReadStarted = yield* Deferred.make<void>();
-      const releaseFirstRead = yield* Deferred.make<void>();
-      const readyHealth = engineHealth("ready", 1);
-      const stoppingHealth = engineHealth("stopping", 1);
-      const health = AtomRef.make(healthFromEngine(readyHealth));
-      let readCount = 0;
-      const healthReads = [
-        Effect.gen(function* () {
-          yield* Deferred.succeed(firstReadStarted, undefined);
-          yield* Deferred.await(releaseFirstRead);
-          return readyHealth;
-        }),
-        Effect.succeed(stoppingHealth),
-      ];
-      const engine = {
-        health: () => {
-          const nextRead = healthReads[readCount] ?? Effect.succeed(stoppingHealth);
-          return Effect.suspend(() =>
-            Effect.sync(() => {
-              readCount += 1;
-            }).pipe(Effect.andThen(nextRead)),
-          );
-        },
-      };
-
-      const staleRefresh = yield* readHealth(engine, health).pipe(Effect.forkDetach);
-      yield* Deferred.await(firstReadStarted);
-
-      const closed = yield* readHealth(engine, health);
-      expect(closed.status).toBe("stopping");
-
-      yield* Deferred.succeed(releaseFirstRead, undefined);
-      yield* Fiber.join(staleRefresh);
-      expect(health.value.status).toBe("stopping");
-    }),
-  );
-
-  it.effect("coalesces concurrent health reads while an active same-epoch read is running", () =>
-    Effect.gen(function* () {
-      const readStarted = yield* Deferred.make<void>();
-      const releaseRead = yield* Deferred.make<void>();
-      let readCount = 0;
-      const coalescedHealth = makeCoalescedHealthReader(
-        () =>
-          Effect.gen(function* () {
-            readCount += 1;
-            yield* Deferred.succeed(readStarted, undefined);
-            yield* Deferred.await(releaseRead);
-            return healthFromEngine(engineHealth("ready", readCount));
-          }),
-        () => 0,
-      );
-
-      const first = yield* coalescedHealth().pipe(Effect.forkChild({ startImmediately: true }));
-      yield* Deferred.await(readStarted);
-      const second = yield* coalescedHealth().pipe(Effect.forkChild({ startImmediately: true }));
-      yield* Deferred.succeed(releaseRead, undefined);
-      const [firstHealth, secondHealth] = yield* Effect.all(
-        [Fiber.join(first), Fiber.join(second)],
-        {
-          concurrency: 2,
-        },
-      ).pipe(Effect.timeout("1 second"));
-
-      expect(readCount).toBe(1);
-      expect(firstHealth.engine.topics.orders.rowCount).toBe(1);
-      expect(secondHealth.engine.topics.orders.rowCount).toBe(1);
-      const thirdHealth = yield* coalescedHealth();
-      expect(readCount).toBe(2);
-      expect(thirdHealth.engine.topics.orders.rowCount).toBe(2);
-    }),
-  );
-
-  it.effect("clears the active health read after a failed read so the next read retries", () =>
-    Effect.gen(function* () {
-      let readCount = 0;
-      const healthReads = [
-        Effect.fail("boom"),
-        Effect.succeed(healthFromEngine(engineHealth("ready", 2))),
-      ];
-      const coalescedHealth = makeCoalescedHealthReader(() =>
-        Effect.gen(function* () {
-          const nextRead =
-            healthReads[readCount] ?? Effect.succeed(healthFromEngine(engineHealth("ready", 3)));
-          readCount += 1;
-          return yield* nextRead;
-        }),
-      );
-
-      const failedHealth = yield* Effect.flip(coalescedHealth());
-      expect(failedHealth).toBe("boom");
-      const recoveredHealth = yield* coalescedHealth();
-      expect(readCount).toBe(2);
-      expect(recoveredHealth.engine.topics.orders.rowCount).toBe(2);
-    }),
-  );
-
-  it.effect("does not strand followers when the active health reader is interrupted", () =>
-    Effect.gen(function* () {
-      const firstReadStarted = yield* Deferred.make<void>();
-      const releaseFirstRead = yield* Deferred.make<void>();
-      const followerJoinedActiveRead = yield* Deferred.make<void>();
-      let epochCheckCount = 0;
-      let readCount = 0;
-      const coalescedHealth = makeCoalescedHealthReader(
-        () =>
-          Effect.gen(function* () {
-            readCount += 1;
-            yield* Deferred.succeed(firstReadStarted, undefined);
-            yield* Deferred.await(releaseFirstRead);
-            return healthFromEngine(engineHealth("ready", readCount));
-          }),
-        () => {
-          epochCheckCount += 1;
-          if (epochCheckCount === 2) {
-            Deferred.doneUnsafe(followerJoinedActiveRead, Effect.void);
-          }
-          return 0;
-        },
-      );
-
-      const leader = yield* coalescedHealth().pipe(Effect.forkChild({ startImmediately: true }));
-      yield* Deferred.await(firstReadStarted);
-      const follower = yield* coalescedHealth().pipe(Effect.forkChild({ startImmediately: true }));
-      yield* Deferred.await(followerJoinedActiveRead).pipe(
-        Effect.timeout("1 second"),
-        Effect.onError(() => Deferred.succeed(releaseFirstRead, undefined).pipe(Effect.asVoid)),
-      );
-
-      const interruptStarted = yield* Deferred.make<void>();
-      const interruptLeader = yield* Effect.gen(function* () {
-        yield* Deferred.succeed(interruptStarted, undefined);
-        yield* Fiber.interrupt(leader);
-      }).pipe(Effect.forkChild({ startImmediately: true }));
-      yield* Deferred.await(interruptStarted);
-      yield* Deferred.succeed(releaseFirstRead, undefined);
-      yield* Fiber.join(interruptLeader);
-      const followerHealth = yield* Fiber.join(follower);
-      const nextHealth = yield* coalescedHealth();
-      expect(readCount).toBe(2);
-      expect(followerHealth.engine.topics.orders.rowCount).toBe(1);
-      expect(nextHealth.engine.topics.orders.rowCount).toBe(2);
-    }),
-  );
-
-  it.effect("starts a fresh health read after the caller epoch changes", () =>
-    Effect.gen(function* () {
-      const firstReadStarted = yield* Deferred.make<void>();
-      const releaseFirstRead = yield* Deferred.make<void>();
-      let epoch = 0;
-      let readCount = 0;
-      const healthReads = [
-        Effect.gen(function* () {
-          yield* Deferred.succeed(firstReadStarted, undefined);
-          yield* Deferred.await(releaseFirstRead);
-          return healthFromEngine(engineHealth("ready", 1));
-        }),
-        Effect.succeed(healthFromEngine(engineHealth("ready", 2))),
-      ];
-      const coalescedHealth = makeCoalescedHealthReader(
-        () =>
-          Effect.gen(function* () {
-            const nextRead =
-              healthReads[readCount] ?? Effect.succeed(healthFromEngine(engineHealth("ready", 3)));
-            readCount += 1;
-            return yield* nextRead;
-          }),
-        () => epoch,
-      );
-
-      const staleHealthRead = yield* coalescedHealth().pipe(
-        Effect.forkChild({ startImmediately: true }),
-      );
-      yield* Deferred.await(firstReadStarted);
-      epoch += 1;
-      const freshHealth = yield* coalescedHealth();
-      yield* Deferred.succeed(releaseFirstRead, undefined);
-      const staleHealth = yield* Fiber.join(staleHealthRead);
-
-      expect(readCount).toBe(2);
-      expect(freshHealth.engine.topics.orders.rowCount).toBe(2);
-      expect(staleHealth.engine.topics.orders.rowCount).toBe(1);
-    }),
-  );
-
-  it.effect("does not let obsolete epoch health reads overwrite fresher cached health", () =>
-    Effect.gen(function* () {
-      const firstReadStarted = yield* Deferred.make<void>();
-      const releaseFirstRead = yield* Deferred.make<void>();
-      const health = AtomRef.make(healthFromEngine(engineHealth("ready", 0)));
-      let epoch = 0;
-      let readCount = 0;
-      const engineHealthReads = [
-        Effect.gen(function* () {
-          yield* Deferred.succeed(firstReadStarted, undefined);
-          yield* Deferred.await(releaseFirstRead);
-          return engineHealth("ready", 1);
-        }),
-        Effect.succeed(engineHealth("ready", 2)),
-      ];
-      const engine = {
-        health: () => {
-          const nextRead = engineHealthReads[readCount] ?? Effect.succeed(engineHealth("ready", 3));
-          return Effect.suspend(() =>
-            Effect.sync(() => {
-              readCount += 1;
-            }).pipe(Effect.andThen(nextRead)),
-          );
-        },
-      };
-      const coalescedHealth = makeCoalescedHealthReader(
-        (readEpoch) =>
-          readHealth(engine, health, {
-            shouldInstall: () => epoch === readEpoch,
-          }),
-        () => epoch,
-      );
-
-      const obsoleteHealthRead = yield* coalescedHealth().pipe(
-        Effect.forkChild({ startImmediately: true }),
-      );
-      yield* Deferred.await(firstReadStarted);
-      epoch += 1;
-      const freshHealth = yield* coalescedHealth();
-      yield* Deferred.succeed(releaseFirstRead, undefined);
-      const obsoleteHealth = yield* Fiber.join(obsoleteHealthRead);
-
-      expect(freshHealth.engine.topics.orders.rowCount).toBe(2);
-      expect(obsoleteHealth.engine.topics.orders.rowCount).toBe(2);
-      expect(health.value.engine.topics.orders.rowCount).toBe(2);
-    }),
-  );
-
-  it.effect("does not let older scheduled health reads overwrite newer installed health", () =>
-    Effect.gen(function* () {
-      const scheduledReadStarted = yield* Deferred.make<void>();
-      const releaseScheduledRead = yield* Deferred.make<void>();
-      const scheduledReadFinished = yield* Deferred.make<void>();
-      const health = AtomRef.make(healthFromEngine(engineHealth("ready", 0)));
-      let installEpoch = 0;
-      let readCount = 0;
-      const engineHealthReads = [
-        Effect.gen(function* () {
-          yield* Deferred.succeed(scheduledReadStarted, undefined);
-          yield* Deferred.await(releaseScheduledRead);
-          return engineHealth("ready", 1);
-        }),
-        Effect.succeed(engineHealth("ready", 2)),
-      ];
-      const engine = {
-        health: () => {
-          const nextRead = engineHealthReads[readCount] ?? Effect.succeed(engineHealth("ready", 3));
-          return Effect.suspend(() =>
-            Effect.sync(() => {
-              readCount += 1;
-            }).pipe(Effect.andThen(nextRead)),
-          );
-        },
-      };
-      const scheduler = yield* makeHealthRefreshScheduler(
-        Effect.gen(function* () {
-          const readInstallEpoch = installEpoch;
-          yield* readHealth(engine, health, {
-            shouldInstall: () => installEpoch === readInstallEpoch,
-            onInstall: () => {
-              installEpoch += 1;
-            },
-          });
-          yield* Deferred.succeed(scheduledReadFinished, undefined);
-        }),
-        "0 millis",
-      );
-
-      yield* scheduler.request;
-      yield* Deferred.await(scheduledReadStarted).pipe(Effect.timeout("1 second"));
-      const freshHealth = yield* readHealth(engine, health, {
-        onInstall: () => {
-          installEpoch += 1;
-        },
-      });
-      yield* Deferred.succeed(releaseScheduledRead, undefined);
-      yield* Deferred.await(scheduledReadFinished).pipe(Effect.timeout("1 second"));
-
-      expect(freshHealth.engine.topics.orders.rowCount).toBe(2);
-      expect(health.value.engine.topics.orders.rowCount).toBe(2);
-      yield* scheduler.close;
-    }),
-  );
-
-  it.effect("lets scheduled health refreshes install and follow up while requests continue", () =>
-    Effect.gen(function* () {
-      const firstReadStarted = yield* Deferred.make<void>();
-      const releaseFirstRead = yield* Deferred.make<void>();
-      const secondReadStarted = yield* Deferred.make<void>();
-      const releaseSecondRead = yield* Deferred.make<void>();
-      const refreshCompleted = yield* Queue.unbounded<void>();
-      const health = AtomRef.make(healthFromEngine(engineHealth("ready", 0)));
-      let installEpoch = 0;
-      let readCount = 0;
-      const engineHealthReads = [
-        Effect.gen(function* () {
-          yield* Deferred.succeed(firstReadStarted, undefined);
-          yield* Deferred.await(releaseFirstRead);
-          return engineHealth("ready", 1);
-        }),
-        Effect.gen(function* () {
-          yield* Deferred.succeed(secondReadStarted, undefined);
-          yield* Deferred.await(releaseSecondRead);
-          return engineHealth("ready", 2);
-        }),
-      ];
-      const engine = {
-        health: () => {
-          const nextRead = engineHealthReads[readCount] ?? Effect.succeed(engineHealth("ready", 3));
-          return Effect.suspend(() =>
-            Effect.sync(() => {
-              readCount += 1;
-            }).pipe(Effect.andThen(nextRead)),
-          );
-        },
-      };
-      const scheduler = yield* makeHealthRefreshScheduler(
-        Effect.gen(function* () {
-          const readInstallEpoch = installEpoch;
-          yield* readHealth(engine, health, {
-            shouldInstall: () => installEpoch === readInstallEpoch,
-            onInstall: () => {
-              installEpoch += 1;
-            },
-          });
-          yield* Queue.offer(refreshCompleted, undefined);
-        }),
-        "0 millis",
-      );
-
-      yield* scheduler.request;
-      yield* Deferred.await(firstReadStarted).pipe(Effect.timeout("1 second"));
-      yield* scheduler.request;
-      yield* Deferred.succeed(releaseFirstRead, undefined);
-      yield* Queue.take(refreshCompleted).pipe(Effect.timeout("1 second"));
-      yield* Deferred.await(secondReadStarted).pipe(Effect.timeout("1 second"));
-      expect(health.value.engine.topics.orders.rowCount).toBe(1);
-      yield* Deferred.succeed(releaseSecondRead, undefined);
-      yield* Queue.take(refreshCompleted).pipe(Effect.timeout("1 second"));
-
-      expect(readCount).toBe(2);
-      expect(health.value.engine.topics.orders.rowCount).toBe(2);
-      yield* scheduler.close;
-    }),
-  );
-
-  it.effect("maps engine errors into runtime errors", () =>
-    Effect.gen(function* () {
-      const runtimeCore = yield* makeViewServerRuntimeCore(viewServer, {});
-      yield* runtimeCore.client.publish("orders", order("a", 10));
-
-      const invalidTopic = yield* Effect.flip(
-        // @ts-expect-error hostile runtime callers can still send unknown topics.
-        runtimeCore.client.publish("missing", order("b", 20)),
-      );
-      const invalidRow = yield* Effect.flip(
-        runtimeCore.client.publish("orders", {
-          id: "bad",
-          customerId: "customer-bad",
-          // @ts-expect-error hostile runtime callers can still send malformed rows.
-          status: "unknown",
-          price: 20,
-          region: "usa",
-          updatedAt: 20,
-        }),
-      );
-      const groupedQuery = yield* runtimeCore.client.snapshot("orders", {
-        groupBy: ["status"],
-        aggregates: { rowCount: { aggFunc: "count" } },
-      });
-      const invalidQuery = yield* Effect.flip(
-        runtimeCore.client.snapshot("orders", {
-          // @ts-expect-error hostile runtime callers can still send unknown projected fields.
-          select: ["prcie"],
-        }),
-      );
-
-      yield* runtimeCore.close;
-      const runtimeUnavailable = yield* Effect.flip(
-        runtimeCore.client.publish("orders", order("closed", 30)),
-      );
-
-      expect(invalidTopic.code).toBe("InvalidTopic");
-      expect(invalidRow.code).toBe("InvalidRow");
-      expect(groupedQuery.rows).toStrictEqual([{ status: "open", rowCount: 1n }]);
-      expect(invalidQuery.code).toBe("InvalidQuery");
-      expect(runtimeUnavailable.code).toBe("RuntimeUnavailable");
-    }),
-  );
-
-  it.effect(
-    "queues a trailing health scheduler refresh when requested while refresh is pending",
-    () =>
-      Effect.gen(function* () {
-        const firstStarted = yield* Deferred.make<void>();
-        const firstFinished = yield* Deferred.make<void>();
-        const secondStarted = yield* Deferred.make<void>();
-        const secondFinished = yield* Deferred.make<void>();
-        let refreshCount = 0;
-        const refreshSteps = [
-          { started: firstStarted, finished: firstFinished },
-          { started: secondStarted, finished: secondFinished },
-        ] as const;
-
-        const scheduler = yield* makeHealthRefreshScheduler(
-          Effect.gen(function* () {
-            const refreshStep = refreshSteps[refreshCount] ?? refreshSteps[1];
-            yield* Effect.sync(() => {
-              refreshCount += 1;
-            });
-            yield* Deferred.succeed(refreshStep.started, undefined);
-            yield* Deferred.await(refreshStep.finished);
-          }),
-          "0 millis",
-        );
-
-        yield* scheduler.request;
-        yield* Deferred.await(firstStarted);
-
-        yield* scheduler.request;
-        yield* Deferred.succeed(firstFinished, undefined);
-        yield* Deferred.await(secondStarted);
-
-        expect(refreshCount).toBe(2);
-        yield* Deferred.succeed(secondFinished, undefined);
-        yield* scheduler.close;
-      }),
-  );
-
-  it.effect("closes a sleeping health scheduler refresh fiber", () =>
-    Effect.gen(function* () {
-      let refreshCount = 0;
-      const scheduler = yield* makeHealthRefreshScheduler(
-        Effect.sync(() => {
-          refreshCount += 1;
-        }),
-        "1 minute",
-      );
-
-      yield* scheduler.request;
-      yield* scheduler.close;
-
-      expect(refreshCount).toBe(0);
-      yield* scheduler.close;
-    }),
-  );
-
-  it.effect("clears active health scheduler state when a refresh interrupts itself", () =>
-    Effect.gen(function* () {
-      const firstStarted = yield* Deferred.make<void>();
-      const secondStarted = yield* Deferred.make<void>();
-      const startedSignals = [firstStarted, secondStarted] as const;
-      let refreshCount = 0;
-      const scheduler = yield* makeHealthRefreshScheduler(
-        Effect.gen(function* () {
-          const started = startedSignals[refreshCount] ?? secondStarted;
-          yield* Effect.sync(() => {
-            refreshCount += 1;
-          });
-          yield* Deferred.succeed(started, undefined);
-          return yield* Effect.interrupt;
-        }),
-        "0 millis",
-      );
-
-      yield* scheduler.request;
-      yield* Deferred.await(firstStarted);
-      yield* Effect.yieldNow;
-      yield* scheduler.request;
-      yield* Deferred.await(secondStarted).pipe(Effect.timeout("1 second"));
-
-      expect(refreshCount).toBe(2);
-      yield* scheduler.close;
-    }),
-  );
-
-  it.effect("ignores health scheduler refresh requests after close", () =>
-    Effect.gen(function* () {
-      let refreshCount = 0;
-      const scheduler = yield* makeHealthRefreshScheduler(
-        Effect.sync(() => {
-          refreshCount += 1;
-        }),
-        "0 millis",
-      );
-
-      yield* scheduler.close;
-      yield* scheduler.request;
-
-      expect(refreshCount).toBe(0);
-    }),
-  );
-});
