+++ packages/runtime-core/src/index.ts
@@ -1,22 +1,16 @@
-import type {
-  DecodableTopicDefinitions,
-  GroupedIncrementalAdmissionLimits,
-} from "@effect-view-server/column-live-view-engine";
+import type { DecodableTopicDefinitions } from "@effect-view-server/column-live-view-engine";
 import type {
   ViewServerConfig,
   GrpcRuntimeClients,
   RuntimeRegions,
-  ViewServerHealth,
   ViewServerRuntimeError,
 } from "@effect-view-server/config";
-import { Effect, type Duration } from "effect";
-import { type RuntimeCoreHealthOverlay, type RuntimeCoreTransportHealth } from "./health";
+import { Effect } from "effect";
 import { makeViewServerRuntimeCoreInternal } from "./internal";
 import type {
-  ViewServerRuntimeCorePublicClient,
-  ViewServerRuntimeCorePublicLiveClient,
-  ViewServerRuntimeCoreServerLiveClient,
-} from "./public-client";
+  ViewServerRuntimeCoreInstance,
+  ViewServerRuntimeCoreOptionsFor,
+} from "./runtime-core-types";
 
 export type { DecodableTopicDefinitions } from "@effect-view-server/column-live-view-engine";
 export type { GroupedIncrementalAdmissionLimits } from "@effect-view-server/column-live-view-engine";
@@ -27,31 +21,11 @@ export type {
   ViewServerRuntimeCorePublicLiveClient,
   ViewServerRuntimeCoreServerLiveClient,
 } from "./public-client";
-
-export type ViewServerRuntimeCoreInstance<Topics extends DecodableTopicDefinitions> = {
-  readonly client: ViewServerRuntimeCorePublicClient<Topics>;
-  readonly liveClient: ViewServerRuntimeCorePublicLiveClient<Topics>;
-  readonly serverLiveClient: ViewServerRuntimeCoreServerLiveClient<Topics>;
-  readonly close: Effect.Effect<void>;
-  readonly requestHealthRefresh: Effect.Effect<void>;
-  readonly refreshHealth: Effect.Effect<ViewServerHealth<Topics>, ViewServerRuntimeError>;
-};
-
-export type ViewServerRuntimeCoreOptions = {
-  readonly groupedIncrementalAdmissionLimits?: Partial<GroupedIncrementalAdmissionLimits>;
-  readonly subscriptionQueueCapacity?: number;
-  readonly transportHealth?: RuntimeCoreTransportHealth<DecodableTopicDefinitions>;
-  readonly healthOverlay?: RuntimeCoreHealthOverlay<DecodableTopicDefinitions>;
-  readonly healthRefreshCadence?: Duration.Input;
-};
-
-export type ViewServerRuntimeCoreOptionsFor<Topics extends DecodableTopicDefinitions> = {
-  readonly groupedIncrementalAdmissionLimits?: Partial<GroupedIncrementalAdmissionLimits>;
-  readonly subscriptionQueueCapacity?: number;
-  readonly transportHealth?: RuntimeCoreTransportHealth<Topics>;
-  readonly healthOverlay?: RuntimeCoreHealthOverlay<Topics>;
-  readonly healthRefreshCadence?: Duration.Input;
-};
+export type {
+  ViewServerRuntimeCoreInstance,
+  ViewServerRuntimeCoreOptions,
+  ViewServerRuntimeCoreOptionsFor,
+} from "./runtime-core-types";
 
 export const makeViewServerRuntimeCore: <
   const Topics extends DecodableTopicDefinitions,
