+++ packages/runtime-core/src/runtime-client.ts
@@ -7,18 +7,15 @@ import type {
   LiveQueryResult,
   RawQuery,
   TopicRow,
-  ViewServerHealth,
   ViewServerTopicConfig,
   ViewServerRuntimeClient,
   ViewServerRuntimeError,
 } from "@effect-view-server/config";
 import { validateLiveQuerySourceRoute } from "@effect-view-server/config";
-import { Effect, type Duration } from "effect";
-import type { AtomRef } from "effect/unstable/reactivity";
+import { Effect } from "effect";
 import {
-  makeHealthRefreshScheduler,
   makeCoalescedHealthReader,
-  readHealth,
+  readHealthSnapshot,
   type RuntimeCoreHealthOverlay,
   type RuntimeCoreTransportHealth,
 } from "./health";
@@ -32,9 +29,6 @@ import { makeSourceOwnershipPolicy } from "./source-ownership-policy";
 export type RuntimeCoreClientInstance<Topics extends DecodableTopicDefinitions> = {
   readonly client: ViewServerRuntimeClient<Topics>;
   readonly internalClient: ViewServerRuntimeCoreInternalClient<Topics>;
-  readonly close: Effect.Effect<void>;
-  readonly requestHealthRefresh: Effect.Effect<void>;
-  readonly refreshHealth: Effect.Effect<ViewServerHealth<Topics>, ViewServerRuntimeError>;
 };
 
 export type ViewServerRuntimeCoreInternalClient<Topics extends DecodableTopicDefinitions> =
@@ -44,68 +38,24 @@ export const makeRuntimeCoreClient = Effect.fn("ViewServerRuntimeCore.client.mak
   <const Topics extends DecodableTopicDefinitions>(
     config: ViewServerTopicConfig<Topics>,
     engine: ColumnLiveViewEngineInternal<Topics>,
-    health: AtomRef.AtomRef<ViewServerHealth<Topics>>,
     runtimeStartedAtNanos: bigint,
     transportHealth: RuntimeCoreTransportHealth<Topics>,
+    requestPushedHealthRefresh: Effect.Effect<void>,
     healthOverlay?: RuntimeCoreHealthOverlay<Topics>,
-    healthRefreshCadence?: Duration.Input,
   ): Effect.Effect<RuntimeCoreClientInstance<Topics>> =>
-    Effect.gen(function* () {
+    Effect.sync(() => {
       const sourceOwnership = makeSourceOwnershipPolicy(config);
-      let healthReadEpoch = 0;
-      let healthInstallEpoch = 0;
-      const bumpHealthReadEpoch = Effect.sync(() => {
-        healthReadEpoch += 1;
-      });
-      const readRuntimeHealth = (epoch: number, installMode: "strict" | "scheduled") => {
-        const installEpoch = healthInstallEpoch;
-        return readHealth(engine, health, {
+      const readFreshRuntimeHealth = () =>
+        readHealthSnapshot(engine, {
           runtimeStartedAtNanos,
           transportHealth,
-          ...(healthOverlay === undefined ? {} : { healthOverlay }),
-          shouldInstall: () =>
-            healthInstallEpoch === installEpoch &&
-            (installMode === "scheduled" || healthReadEpoch === epoch),
-          onInstall: () => {
-            healthInstallEpoch += 1;
-          },
+          healthOverlay,
         });
-      };
-      const healthReader = makeCoalescedHealthReader(
-        (epoch) => readRuntimeHealth(epoch, "strict"),
-        () => healthReadEpoch,
-      );
-      const scheduledHealthReader = makeCoalescedHealthReader(
-        (epoch) => readRuntimeHealth(epoch, "scheduled"),
-        () => healthReadEpoch,
-      );
-      const scheduledHealthRefresh = Effect.fn(
-        "ViewServerRuntimeCore.client.healthRefresh.scheduled",
-      )(function* () {
-        yield* scheduledHealthReader();
-      });
-      const healthRefreshScheduler = yield* makeHealthRefreshScheduler(
-        scheduledHealthRefresh(),
-        healthRefreshCadence,
-      );
-      const requestHealthRefresh = Effect.fn("ViewServerRuntimeCore.client.healthRefresh.request")(
-        function* () {
-          yield* Effect.uninterruptible(
-            bumpHealthReadEpoch.pipe(Effect.andThen(healthRefreshScheduler.request)),
-          );
-        },
-      );
-      const refreshHealthNow = Effect.fn("ViewServerRuntimeCore.client.healthRefresh.now")(
-        function* () {
-          return yield* Effect.uninterruptible(
-            bumpHealthReadEpoch.pipe(Effect.andThen(healthReader())),
-          );
-        },
-      );
+      const healthReader = makeCoalescedHealthReader(readFreshRuntimeHealth);
       const mutationPipeline = makeRuntimeCoreMutationPipeline(
         config,
         engine,
-        requestHealthRefresh(),
+        requestPushedHealthRefresh,
       );
       const snapshot = <
         Topic extends Extract<keyof Topics, string>,
@@ -143,9 +93,6 @@ export const makeRuntimeCoreClient = Effect.fn("ViewServerRuntimeCore.client.mak
           health: internalClient.health,
         },
         internalClient,
-        close: healthRefreshScheduler.close,
-        requestHealthRefresh: requestHealthRefresh(),
-        refreshHealth: refreshHealthNow(),
       } satisfies RuntimeCoreClientInstance<Topics>;
     }),
 );
