+++ packages/runtime-core/vite.config.ts
@@ -13,7 +13,7 @@ export default defineConfig({
     coverage: {
       provider: "istanbul",
       include: ["src/**/*.ts"],
-      exclude: ["src/**/*.test.ts", "src/**/*.test-d.ts"],
+      exclude: ["src/**/*.test.ts", "src/**/*.test-d.ts", "src/test-support/**/*.ts"],
       reporter: ["text"],
       thresholds: {
         "100": true,
